requirements.txt
tqdm
python-dotenv
Django
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
psycopg2-binary
uvicorn